AI_server/requirements.txt
pyyaml>=5.1
tensorboard
pycocotools
pyclipper
Polygon3
colorlog
ocrd-fork-pylsd
einops
prefetch_generator
imgaug
lmdb
transformers
torchnet
webcolors
pandas_profiling
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
opencv-python==4.8.1.78
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.5.0
numpy>=1.24.0
pandas>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
scipy>=1.10.0
